requirements.txt
langchain
langchain-community
watchdog
streamlit
gitpython
streamlit-chat
streamlit-extras
google-generativeai
pinecone-client
openai==0.28
tiktoken
google-cloud-aiplatform
black
python-dotenv
google-search-results
tweepy
google-api-python-client
google-auth-oauthlib
wikipedia
faker
google-cloud-discoveryengine
google-auth-httplib2
numexpr 
